+++ CHANGELOG.md
@@ -25,6 +25,9 @@ The present file will list all changes made to the project; according to the
 - Format of `Message-Id` header sent in Tickets notifications changed to match format used by other items.
 - Added `DB::truncate()` to replace raw SQL queries
 - Impact context `positions` field type changed from `TEXT` to `MEDIUMTEXT`
+- Field `date` of KnowbaseItem has been renamed to `date_creation`.
+- Field `date` of Notepad has been renamed to `date_creation`.
+- Field `date` of ProjectTask has been renamed to `date_creation`.
 
 #### Deprecated
 - Usage of `GLPI_FORCE_EMPTY_SQL_MODE` constant